+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added
   - `--only-changed` option
   - Colored file output
-  - New mypy component
+  - New components:
+    - mypy
+    - git check-ignore
 - Changed
   - new python wrapper
   - Optimize reading list of files